- fix deprecation

+++ classes/template/Template.php
@@ -509,9 +509,10 @@
 			string $comment, int $width = 120, string $char = '-'
 		)
 		: string {
-			$w = $width - 14; // 7 for start comment, 5 for end comment + 2 spaces before and after comment
+   
+			$w = (float) $width - 14; // 7 for start comment, 5 for end comment + 2 spaces before and after comment
             $w1 = 2/3 * $w;
-			if ( strlen( $comment ) > $w1 ) {
+			if ( (float) strlen( $comment ) > (float) $w1 ) {
 				$comment = '[...]' . substr( $comment, $w1 );
 			}
 			$s = str_repeat( $char,intval(( $w - strlen( $comment ) ) / 2) );
